requirements.txt
openai-whisper
moviepy
Pillow
googletrans==4.0.0rc1
